requirements.txt
fastapi
langchain
uvicorn
python-multipart
python-dotenv
PyPDF2
pdf2image
langchain
langchain-core
langchain-groq
langchain-text-splitters
google
google-api-core
google-auth
google-cloud
google-cloud-vision
googleapis-common-protos
googletrans